mapper y pacientedto reestructurado

+++ back-end/back/src/main/java/com/remind/back/services/PacienteServiceImpl.java
@@ -3,15 +3,18 @@ package com.remind.back.services;
 import java.util.List;
 import java.util.Optional;
 
-import com.remind.back.dto.PacienteDTO;
+import com.remind.back.Mapper.PacienteMapper;
+import com.remind.back.dto.PacienteInputDTO;
+import com.remind.back.dto.PacienteOutputDTO;
 import com.remind.back.entities.Paciente;
-import com.remind.back.entities.PacienteTerapeuta;
-import com.remind.back.entities.Terapeuta;
+
 import com.remind.back.repositories.PacienteRepository;
 
 import org.springframework.transaction.annotation.Transactional;
+import java.util.NoSuchElementException;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -20,57 +23,75 @@ public class PacienteServiceImpl implements PacienteService {
     @Autowired
     private PacienteRepository pacienteRepository;
 
+    @Autowired
+    private PacienteMapper pacienteMapper;
+
     @Override
     @Transactional
-    public Paciente createPaciente(PacienteDTO pacienteDTO) {
-        Paciente paciente = new Paciente();
-        paciente.setNombre(pacienteDTO.getNombre());
-        paciente.setApellido(pacienteDTO.getApellido());
-        paciente.setEmail(pacienteDTO.getEmail());
-        paciente.setContraseña(pacienteDTO.getContraseña());
-        paciente.setTelefono(pacienteDTO.getTelefono());
-        paciente.setEnfermedad(pacienteDTO.getEnfermedad());
-        paciente.setEdad(pacienteDTO.getEdad());
-        paciente.setNombreResponsable(pacienteDTO.getNombreResponsable());
-        PacienteTerapeuta pacienteTerapeuta = new PacienteTerapeuta();
-        pacienteTerapeuta.setTerapeuta(pacienteDTO.getTerapeutaId());
-        pacienteRepository.save(paciente);
-        pacienteTerapeuta.setPaciente(paciente);
-        return paciente;
+    public PacienteOutputDTO createPaciente(PacienteInputDTO pacienteInputDTO) {
+    
+        Paciente paciente = pacienteMapper.PacienteInputDTOToPaciente(pacienteInputDTO);
+        return pacienteMapper.PacienteToPacienteOutputDTO(pacienteRepository.save(paciente));
     }
 
     @Override
-    public List<Paciente> getAllPacientes() {
-        return pacienteRepository.findAll();
+    public List<PacienteOutputDTO> getAllPacientes(int page,int size) {
+        List<Paciente> pacientes = pacienteRepository.findAll(PageRequest.of(page,size)).getContent();
+        return pacienteRepository.findAll(PageRequest.of(page,size))
+            .getContent()
+            .stream()
+            .map(pacienteMapper::PacienteToPacienteOutputDTO)
+            .toList();
     }
 
     @Override
     @Transactional(readOnly = true)
-    public Optional<Paciente> getPacienteById(Integer id) {
-        return pacienteRepository.findById(id);
+    public PacienteOutputDTO getPacienteById(Integer id) {
+        Paciente paciente = pacienteRepository.findById(id)
+            .orElseThrow(() ->  new NoSuchElementException("El paciente con id " + id + " no existe"));
+        return pacienteMapper.PacienteToPacienteOutputDTO(paciente);
     }
 
     @Override
     @Transactional
     public void deletePaciente(Integer id) {
+        if(!pacienteRepository.existsById(id)){
+            throw new NoSuchElementException("No existe un paciente con ese id y no se puede eliminar");
+        }
         pacienteRepository.deleteById(id);
     }
 
     @Override
     @Transactional
-    public void updatePaciente(Integer id, PacienteDTO pacienteDTO) {
-        Paciente paciente = pacienteRepository.findById(id).get();
-        paciente.setNombre(pacienteDTO.getNombre());
-        paciente.setApellido(pacienteDTO.getApellido());
-        paciente.setEmail(pacienteDTO.getEmail());
-        paciente.setContraseña(pacienteDTO.getContraseña());
-        paciente.setTelefono(pacienteDTO.getTelefono());
-        paciente.setEnfermedad(pacienteDTO.getEnfermedad());
-        paciente.setEdad(pacienteDTO.getEdad());
-        paciente.setNombreResponsable(pacienteDTO.getNombreResponsable());
-        // parte de modificar
+    public PacienteOutputDTO updatePaciente(Integer id, PacienteInputDTO pacienteDTO) {
         
-        pacienteRepository.save(paciente);
+        Paciente paciente = pacienteRepository.findById(id)
+            .orElseThrow(() ->  new NoSuchElementException("El paciente con id " + id + " no existe"));
+        if(pacienteDTO.getNombre() != null){
+            paciente.setNombre(pacienteDTO.getNombre());
+        }
+        if(pacienteDTO.getEnfermedad() !=null){
+            paciente.setEnfermedad(pacienteDTO.getEnfermedad());
+        }
+        if(pacienteDTO.getApellido() != null){
+            paciente.setApellido(pacienteDTO.getApellido());
+        }
+        if(pacienteDTO.getEdad() != null){
+            paciente.setEdad(pacienteDTO.getEdad());
+        }    
+        if(pacienteDTO.getTelefono() != null){
+            paciente.setTelefono(pacienteDTO.getTelefono());
+        }
+        if(pacienteDTO.getEmail() != null){
+            paciente.setEmail(pacienteDTO.getEmail());
+        }
+        
+        if(pacienteDTO.getFechaNacimiento() != null){
+            paciente.setFechaNacimiento(pacienteDTO.getFechaNacimiento());
+        }
+        
+        return pacienteMapper.PacienteToPacienteOutputDTO(pacienteRepository.save(paciente));
+      
     }
 
 }
